User Interface Elements
This creates the UI of the app which includes the sign in and signup
pages, a dashboard page where the user can view a list of his or her
diary entries, view the content of an entry, and  modify an entry.
and a page  for the user to edit his or her profile details, set
reminders and view the number of entries since registration.